Training NN
In this step, there were some errors. For example, the neural networks cannot use the certain column name, so I had to clean the data again and split it again and restructure the whole data again, which took time. Also, I have re-trained all the modules from notebooks 1, 2, 3, 4, and 5, and then I'm working on the sixth one on the neural networks. It is running, but I intend to branch it out and try a new experiment and a new pipeline that should be working after I am done.